Feature/render serialization

+++ src/shaders/sprite2d.vert
@@ -6,7 +6,7 @@ layout (location = 1) in vec2 aTexCoord;
 uniform mat4 model;
 uniform mat4 projection;
 uniform vec2 size;
-uniform vec2 offset;
+uniform ivec2 offset;
 uniform float z;
 
 
@@ -14,8 +14,8 @@ out vec2 tex_coord;
 
 void main() {
     vec2 pos = vec2(Position.x * size.x, Position.y * size.y);
-    pos.x += offset.x;
-    pos.y += offset.y;
+    pos.x += float(offset.x);
+    pos.y += float(offset.y);
 
     gl_Position = projection * model * vec4(pos.xy, z, 1.0);
     tex_coord = aTexCoord;
